Update UserRepository implementation

+++ src/main/java/com/eyuup/security/repository/UserRepository.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.eyuup.security.user.User;
+import com.eyuup.security.model.User;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User,Long> {
    Optional<User> findByEmail(String email);
 }
